% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{sfs_connect}
\alias{sfs_connect}
\title{Connect to a Synology File Station}
\usage{
sfs_connect(url, user, pass, version = 0L, verbose = sfs_verbose())
}
\arguments{
\item{url}{string: URL to connect to, e.g. "https://my.sfs.com/webapi/"}

\item{user}{string: username. If missing, will be taken from the environment variable `SFS_USER`}

\item{pass}{string: password. If missing, will be taken from the environment variable `SFS_PASS`}

\item{version}{integer: SYNO.Api.Auth version to use. If 0, `sfs_info(url)$SYNO.API.Auth$maxVersion` will be called to determine the appropriate value to use (but adds a second or two to the connection time)}

\item{verbose}{logical: give debug info?}
}
\value{
An sfsclient object
}
\description{
Connect to a Synology File Station
}
\references{
<https://global.download.synology.com/download/Document/Software/DeveloperGuide/Package/FileStation/All/enu/Synology_File_Station_API_Guide.pdf>
}
