% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{sfs_download}
\alias{sfs_download}
\title{Download files or folders}
\usage{
sfs_download(
  sfs,
  path,
  out_dir,
  mode = "download",
  progress = TRUE,
  verbose = sfs_verbose()
)
}
\arguments{
\item{sfs}{sfsclient: as returned by [sfs_connect()]}

\item{path}{character: one or more file or folder paths (each starting with a shared folder. Probably needs to start with a "/")}

\item{out_dir}{string: destination directory. If not provided, a temporary directory will be used}

\item{mode}{string: "download" (default) or "open"}

\item{progress}{logical: show download progress?}

\item{verbose}{logical: give debug info?}
}
\value{
A data.frame or `NULL`
}
\description{
Download files or folders
}
\references{
<https://global.download.synology.com/download/Document/Software/DeveloperGuide/Package/FileStation/All/enu/Synology_File_Station_API_Guide.pdf>
}
