% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{sfs_fs_info}
\alias{sfs_fs_info}
\title{Retrieve File Station information}
\usage{
sfs_fs_info(sfs, verbose = sfs_verbose())
}
\arguments{
\item{sfs}{sfsclient: as returned by [sfs_connect()]}

\item{verbose}{logical: give debug info?}
}
\value{
A list, with information potentially including
* is_manager - if the logged-in user is an administrator
* support_virtual_protocol - types of virtual file system which the logged user is able to mount on (CIFS, NFS, ISO)
* support_sharing - if the logged-in user can share files/folders
* hostname - host name
}
\description{
Retrieve File Station information
}
\references{
<https://global.download.synology.com/download/Document/Software/DeveloperGuide/Package/FileStation/All/enu/Synology_File_Station_API_Guide.pdf>
}
