% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{sfs_info}
\alias{sfs_info}
\title{Retrieve API information from a Synology File Station}
\usage{
sfs_info(url, verbose = sfs_verbose())
}
\arguments{
\item{url}{string: URL to connect to, e.g. "https://my.sfs.com/webapi/"}

\item{verbose}{logical: give debug info?}
}
\value{
An sfsclient object
}
\description{
Retrieve API information from a Synology File Station
}
\references{
<https://global.download.synology.com/download/Document/Software/DeveloperGuide/Package/FileStation/All/enu/Synology_File_Station_API_Guide.pdf>
}
