% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{sfs_list}
\alias{sfs_list}
\title{List files in a shared folder}
\usage{
sfs_list(sfs, path, additional = c("time", "size"), verbose = sfs_verbose())
}
\arguments{
\item{sfs}{sfsclient: as returned by [sfs_connect()]}

\item{path}{string: folder path. Probably needs to start with a "/"}

\item{additional}{character: one or more of "real_path", "size", "owner", "time", "perm", "type", "mount_point_type"}

\item{verbose}{logical: give debug info?}
}
\value{
A data.frame or `NULL`
}
\description{
List files in a shared folder
}
\references{
<https://global.download.synology.com/download/Document/Software/DeveloperGuide/Package/FileStation/All/enu/Synology_File_Station_API_Guide.pdf>
}
