% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{sfs_search_start}
\alias{sfs_search_start}
\alias{sfs_search_list}
\alias{sfs_search_stop}
\alias{sfs_search_clean}
\title{Search files according to given criteria}
\usage{
sfs_search_start(
  sfs,
  path,
  recursive = TRUE,
  pattern = NULL,
  extension = NULL,
  verbose = sfs_verbose()
)

sfs_search_list(
  sfs,
  taskid,
  additional = c("time", "size"),
  clean = TRUE,
  verbose = sfs_verbose()
)

sfs_search_stop(sfs, taskid, verbose = sfs_verbose())

sfs_search_clean(sfs, taskid, verbose = sfs_verbose())
}
\arguments{
\item{sfs}{sfsclient: as returned by [sfs_connect()]}

\item{path}{string: one or more folder paths. Probably needs to start with a "/"}

\item{recursive}{logical: search recursively?}

\item{pattern}{string: search for files whose names and extensions match a case-insensitive glob pattern. If `pattern` doesn't contain any glob syntax ("?" or "*"), "*" will be added at begin and end of the string automatically for partially matching the pattern. You can use " " to separate multiple glob patterns}

\item{extension}{string: Search for files whose extensions match a file type pattern in a case-insensitive glob pattern. If you give this criterion, folders aren't matched. You can use commas "," to separate multiple glob patterns.}

\item{verbose}{logical: give debug info?}

\item{taskid}{string: search task ID as returned by `sfs_search_start`}

\item{additional}{character: one or more of "real_path", "size", "owner", "time", "perm", "type", "mount_point_type"}

\item{clean}{logical: if the search has finished, stop the search process and remove the temporary database? Note that further `sfs_search_list` requests cannot then be issued against this `taskid`}
}
\value{
A data.frame or `NULL`
}
\description{
Search files according to given criteria
}
\references{
<https://global.download.synology.com/download/Document/Software/DeveloperGuide/Package/FileStation/All/enu/Synology_File_Station_API_Guide.pdf>
}
