% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{sfs_verbose}
\alias{sfs_verbose}
\title{Set global sfsclient verbosity}
\usage{
sfs_verbose(verbose)
}
\arguments{
\item{verbose}{logical: give debug info?}
}
\value{
If called with no parameters, the current verbosity setting. Otherwise sets the value and returns it
}
\description{
A convenience function that sets the verbosity level for all `sfsclient` functions, without having to explicitly set it on each function call.
}
\examples{
sfs_verbose(TRUE)
sfs_verbose()

}
